requirements.txt
# Core dependencies - 工具运行必需的库
openai~=1.102.0
ddgs~=9.5.0
aiohttp~=3.12.0
beautifulsoup4~=4.13.0
html2text~=2025.4.15
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0
toml>=0.10.0

# Testing dependencies
pytest~=8.4.1
